Rename kinto_updater into kinto_signer
