+++ .coveragerc
@@ -1,4 +1,4 @@
 [run]
-omit = kinto_updater/tests/*
-       kinto_updater/hook.py
-       kinto_updater/gather_remote.py
+omit = kinto_signer/tests/*
+       kinto_signer/hook.py
+       kinto_signer/gather_remote.py
