+++ kinto.ini
@@ -8,14 +8,14 @@ use = egg:kinto
 kinto.userid_hmac_secret = aujourd'hui il fait froid en bretagne.
 multiauth.policies = basicauth
 
-kinto.includes = kinto_updater.hook
+kinto.includes = kinto_signer.hook
 
 ; Kinto Updater configuration.
 ; XXX Put this in a section instead.
-kinto_updater.bucket =
-kinto_updater.collection =
-kinto_updater.remote_server_url = http://localhost:7777
-kinto_updater.private_key = 'test.pem'
+kinto_signer.bucket =
+kinto_signer.collection =
+kinto_signer.remote_server_url = http://localhost:7777
+kinto_signer.private_key = 'test.pem'
 
 [loggers]
 keys = root, cliquet
