+++ kinto_signer/__init__.py
