+++ kinto_signer/hasher.py
