+++ kinto_signer/signer.py
