+++ kinto_signer/tests/config/ecdsa.public.pem
