+++ kinto_signer/tests/config/remote.ini
