+++ kinto_signer/tests/config/signer.ini
@@ -11,13 +11,13 @@ kinto.flush_endpoint_enabled = true
 
 ; Kinto Updater configuration.
 ; XXX Put this in a section instead.
-kinto.includes = kinto_updater.hook
+kinto.includes = kinto_signer.hook
 
-kinto_updater.bucket = buck
-kinto_updater.collection = coll
-kinto_updater.remote_server_url = http://localhost:7777/v1
-kinto_updater.remote_server_auth = user:p4ssw0rd
-kinto_updater.private_key = kinto_updater/tests/config/ecdsa.private.pem
+kinto_signer.bucket = buck
+kinto_signer.collection = coll
+kinto_signer.remote_server_url = http://localhost:7777/v1
+kinto_signer.remote_server_auth = user:p4ssw0rd
+kinto_signer.private_key = kinto_signer/tests/config/ecdsa.private.pem
 
 [loggers]
 keys = root, cliquet
