+++ kinto_signer/tests/support.py
