+++ kinto_signer/updater.py
