+++ tox.ini
@@ -4,11 +4,11 @@ envlist = py27,flake8
 [testenv:py27]
 commands =
     python --version
-    py.test kinto_updater/tests --cov-report term-missing --cov-fail-under 100 --cov kinto_updater {posargs}
+    py.test kinto_signer/tests --cov-report term-missing --cov-fail-under 100 --cov kinto_signer {posargs}
 deps = -rdev-requirements.txt
 install_command = pip install --pre {opts} {packages}
 
 [testenv:flake8]
-commands = flake8 kinto_updater
+commands = flake8 kinto_signer
 deps =
     flake8
